upd: Refactored DAG 'create-price-table'
